# Conflicts:
#	common/data/src/main/java/org/thingsboard/server/common/data/HasVersion.java

+++ common/data/src/main/java/org/thingsboard/server/common/data/HasVersion.java
@@ -17,7 +17,7 @@ package org.thingsboard.server.common.data;
 
 public interface HasVersion {
 
-    Integer getVersion();
+    Long getVersion();
 
     void setVersion(Integer version);
